requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
typing-extensions>=4.0.0
aiohttp>=3.9.0
pytz>=2023.3
python-dateutil>=2.8.2
json-schema>=0.1.0
requests>=2.31.0
asyncio>=3.4.3

# Local model support
httpx>=0.24.1

# Visualization and monitoring
rich>=13.0.0
tabulate>=0.9.0

# Knowledge management
faiss-cpu>=1.7.4
networkx>=3.0.0